some refactoring
