--- CalorieContent/CalorieContentMain/src/Core/Repositories/Ingredient/IIngredientRepo.cs
@@ -1,9 +1,0 @@
-﻿using CalorieContent.Core.Repositories.Base;
-using CalorieContent.Domain.Entities;
-
-namespace CalorieContent.Core.Repositories
-{
-    public interface IIngredientRepo : IRepository<Ingredient>
-    {
-    }
-}
