--- CalorieContent/CalorieContentMain/src/Core/Repositories/Recipe/IRecipeRepo.cs
@@ -1,9 +1,0 @@
-﻿using CalorieContent.Core.Repositories.Base;
-using CalorieContent.Domain.Entities;
-
-namespace CalorieContent.Core.Repositories
-{
-    public interface IRecipeRepo : IRepository<Recipe>
-    {
-    }
-}
