+++ CalorieContent/CalorieContentMain/src/Repositories/Ingredient/IIngredientRepo.cs
@@ -1,0 +1,8 @@
+﻿using CalorieContent.Repositories.Base;
+
+namespace CalorieContent.Repositories.Ingredient
+{
+    public interface IIngredientRepo : IRepository<Domain.Entities.Ingredient>
+    {
+    }
+}
