+++ CalorieContent/CalorieContentMain/src/Repositories/Recipe/IRecipeRepo.cs
@@ -1,0 +1,8 @@
+﻿using CalorieContent.Repositories.Base;
+
+namespace CalorieContent.Repositories.Recipe
+{
+    public interface IRecipeRepo : IRepository<Domain.Entities.Recipe>
+    {
+    }
+}
